[#140]refactor: ProductLikeButton 사용하도록 수정 및 props 설정 추가
